This breaks the code but we kind of know what it is (Next and SSR).
