+++ package.json
@@ -18,7 +18,6 @@
     "@emotion/styled": "^11.10.6",
     "@hookform/resolvers": "^2.9.11",
     "@next/font": "13.2.1",
-    "@types/circular-dependency-plugin": "^5.0.5",
     "@types/node": "18.14.1",
     "@types/react": "18.0.28",
     "@types/react-dom": "18.0.11",
